+++ src-tauri/permissions/default.toml
@@ -14,6 +14,7 @@ commands.allow = [
   "mfa_finish_code",
   "mfa_poll_openid",
   "mfa_connect_mobile_approve",
+  "mfa_fido2_pin",
   "cancel_mfa",
   "all_locations",
   "has_any_visible_locations",
